use gather operation for cropping
